8 upload file crud user

+++ src/main/java/com/cg/spblaguna/model/dto/res/RoomResDTO.java
@@ -1,10 +1,15 @@
 package com.cg.spblaguna.model.dto.res;
 
 import com.cg.spblaguna.model.*;
+import com.cg.spblaguna.model.enumeration.ERoomType;
+import com.cg.spblaguna.model.enumeration.EStatusRoom;
+import com.cg.spblaguna.model.enumeration.EViewType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -27,11 +32,11 @@ public class RoomResDTO {
 
     private PerType perType;
 
-    private Long pricePerNight;
+    private BigDecimal pricePerNight;
 
-    private Long acreage;
+    private BigDecimal acreage;
 
-    private Long sleep;
+    private Integer sleep;
 
     private String description;
